+++ curvlinops/ekfac.py
@@ -27,6 +27,208 @@ from curvlinops.kfac import (
 from curvlinops.kfac_utils import extract_patches
 
 
+def compute_eigenvalue_correction_linear_weight_sharing(
+    g: Tensor,
+    ggT_eigvecs: Tensor,
+    a: Tensor,
+    aaT_eigvecs: Union[Tensor, None],
+    _force_strategy: Optional[str] = None,
+) -> Tensor:
+    r"""Computes eigenvalue corrections for a linear layer with weight sharing.
+
+    Chooses between two computational strategies depending on memory requirements.
+
+    Args:
+        g: Output gradients of the layer with shape
+            ``[N, S, D1]``, where ``N`` is the batch size, ``S`` the weight sharing
+            dimension, and ``D1`` the output dimension.
+        ggT_eigvecs: Eigenvectors of the gradient covariance with shape
+            ``[D1, D1]``.
+        a: Layer inputs with shape ``[N, S, D2]``, where ``D2`` is the input dimension.
+        aaT_eigvecs: Eigenvectors of the input covariance with shape
+            ``[D2, D2]`` or ``None`` if the layer has no weights (bias only).
+        _force_strategy: If specified, forces the use of either ``'gramian'`` or
+            ``'per_example_gradients'`` strategy. If ``None``, the strategy is chosen
+            based on memory requirements. Defaults to ``None``. This flag serves mainly
+            for testing purposes.
+
+    Returns:
+        The eigencorrection with shape ``[D1, D2]`` (or ``[D1]`` for the bias case).
+
+    Raises:
+        ValueError: If an invalid ``_force_strategy`` is provided.
+
+    Below we explain the mathematical details of what this function does. The mapping is
+    is as follows: (``g``, :math:`\mathbf{Y}`), (``ggT_eigvecs``, :math:`\mathbf{Q}_1`),
+    (``a``, :math:`\mathbf{X}`), (``aaT_eigvecs``, :math:`\mathbf{Q}_2`).
+
+    Note:
+        **Introduction:** In the following, let :math:`D_1` be the output dimension
+        of the layer, :math:`D_2` the input dimension, :math:`S` the weight sharing
+        dimension, and :math:`N` the batch size.
+
+        Given the layer inputs :math:`\mathbf{X} \in \mathbb{R}^{N \times S \times D_2}`,
+        output gradients :math:`\mathbf{Y} \in \mathbb{R}^{N \times S \times D_1}`, and a
+        Kronecker-factored basis :math:`\mathbf{Q}_1 \otimes \mathbf{Q}_2` with factors
+        :math:`\mathbf{Q}_i \in \mathbb{R}^{D_i \times D_i}`, our goal is to compute the
+        eigencorrection :math:`\mathbf{E} \in \mathbb{R}^{D_1 \times D_2}` which has the
+        same shape as the layer's weights.
+
+        The common way to do that is to compute the per-example gradients
+        :math:`\mathbf{G} \in \mathbb{R}^{N \times D_1 \times D_2}` with
+
+        .. math::
+            \mathbf{G}_{n,d_1,d_2}
+            =
+            \sum_s \mathbf{Y}_{n,s,d_1} \mathbf{X}_{n,s,d_2},
+
+        rotate them into the Kronecker-factored basis,
+
+        .. math::
+            \mathbf{\tilde{G}}_{n,d_1,d_2}
+            =
+            \sum_{i,j} \mathbf{G}_{n,i,j} \mathbf{Q}_{1,i,d_1} \mathbf{Q}_{2,j,d_2},
+
+        and compute the correction by squaring and summing out the batch dimension,
+
+        .. math::
+            \mathbf{E}_{d_1,d_2}
+            =
+            \sum_{n} \mathbf{\tilde{G}}_{n,d_1,d_2}^2.
+
+        Building up the per-example gradients can be extremely memory-costly.
+        Therefore, we also consider an alternative approach which can have smaller
+        memory footprint if the weight sharing is mild.
+
+    Note:
+        **(1) Cost analysis of per-example gradient approach:** The peak memory of
+        building up per-example gradients is dominated by :math:`N D_1 D_2`.
+
+        We have two options to compute the rotated per-example gradient.
+
+        1. First compute :math:`\mathbf{G}` and then rotate it. The first step costs
+           :math:`N S D_1 D_2` time and the rotation costs
+           :math:`N D_1 D_2 (D_1 + D_2)` time.
+        2. Rotate the activations and output gradients, then compute the rotated
+           per-example gradient. The rotations cost :math:`N S (D_1^2 + D_2^2)` time
+           The last step is :math:`N S D_1 D_2` time.
+
+        So in practise, we should prefer the first approach over the second if
+
+        .. math::
+            D_1 D_2 (D_1 + D_2) < S (D_1^2 + D_2^2).
+
+        In the implementation, ``opt-einsum`` will automatically do that for us.
+
+        Adding the cost for squaring and contracting, the overall cost is
+        :math:`N S D_1 D_2 + N \min(S (D_1^2 + D_2^2), D_1 D_2 (D_1 + D_2)) + 2 N D_1 D_2`.
+
+    Note:
+        **(2) Cost analysis of Gramian contraction approach:** A way to avoid building
+        up per-example gradients is to write the eigencorrection as big contraction of
+        the rotated activations :math:`\mathbf{\tilde{X}}, \mathbf{\tilde{Y}}` and then
+        rearrange the contractions such that the batch dimension can be directly summed:
+
+        .. math::
+            \mathbf{E}_{d_1,d_2}
+            =
+            \sum_{n}
+            \left(
+            \sum_{s} \mathbf{\tilde{Y}}_{n,s,d_1} \mathbf{\tilde{X}}_{n,s,d_2}
+            \right)
+            \left(
+            \sum_{t} \mathbf{\tilde{Y}}_{n,t,d_1} \mathbf{\tilde{X}}_{n,t,d_2}
+            \right)
+            \\
+            =
+            \sum_{n} \sum_{s} \sum_{t}
+            \left(
+            \mathbf{\tilde{Y}}_{n,s,d_1} \mathbf{\tilde{Y}}_{n,t,d_1}
+            \right)
+            \left(
+            \mathbf{\tilde{X}}_{n,s,d_2} \mathbf{\tilde{X}}_{n,t,d_2}
+            \right)
+
+        This requires building up the Gramians
+
+        .. math::
+            \mathbf{G^Y}_{n,s,t,d_1}
+            =
+            \mathbf{\tilde{Y}}_{n,s,d_1} \mathbf{\tilde{Y}}_{n,t,d_1},
+            \\
+            \mathbf{G^X}_{n,s,t,d_2}
+            =
+            \mathbf{\tilde{X}}_{n,s,d_2} \mathbf{\tilde{X}}_{n,t,d_2}.
+
+        Peak memory is dominated by :math:`N S^2 (D_1 + D_2)`.
+        The time is :math:`N S (D_1^2 + D_2^2)` for the rotations,
+        :math:`N S^2 (D_1 + D_2)` for building up the Gramians, and
+        :math:`N S^2 D_1 D_2` for the final contraction.
+        In total, this is :math:`N S (D_1^2 + D_2^2) + N S^2 (D_1 + D_2 + D_1 D_2)`.
+
+    **We select the approach with the smaller memory footprint**, i.e. Gramian
+    contraction if :math:`S^2 (D_1 + D_2) < D_1 D_2`, and squaring per-example
+    gradients otherwise. So generally speaking, the more weight sharing, the
+    better building up per-example gradients will be.
+
+    In the extreme case :math:`S=1` (no weight sharing), the Gramian
+    contraction approach uses only :math:`N (D_1 + D_2) < N D_1 D_2` memory
+    compared to the per-example gradient approach. In terms of time, the
+    Gramian contraction uses :math:`N (D_1^2 + D_2^2 + D_1 + D_2 + D_1 D_2) < N
+    (3 D_1 D_2 + D_1^2 + D_2^2)` compared to the per-example gradient.
+    """
+    strategies = {"gramian", "per_example_gradients", None}
+    if _force_strategy not in strategies:
+        raise ValueError(
+            f"Invalid _force_strategy: {_force_strategy}. Supported: {strategies}."
+        )
+
+    Q1, Q2 = ggT_eigvecs, aaT_eigvecs
+    Y, X = g, a
+
+    if Q2 is None:  # -> 1d (bias) case
+        eigencorrection = (
+            einsum(Q1, Y, "j d1, batch shared j -> batch d1").square_().sum(0)
+        )
+
+    else:  # -> 2d (weight or weight+bias) case
+        # Determine approach: Gramian contraction or per-example gradients
+        (_, S, D1), (_, _, D2) = g.shape, a.shape
+
+        # Determine approach based on _force_strategy or memory requirements
+        use_gramian = (
+            _force_strategy == "gramian"
+            if _force_strategy is not None
+            # Memory of per-example gradients is dominated by N * D1 * D2
+            # Memory of Gramian contraction is dominated by N * S^2 * (D1 + D2)
+            # We choose the approach that requires less memory.
+            else S**2 * (D1 + D2) < D1 * D2
+        )
+
+        if use_gramian:  # -> Gramian approach
+            X_rot = einsum(X, Q2, "batch shared j, j d2 -> batch shared d2")
+            Y_rot = einsum(Y, Q1, "batch shared i, i d1 -> batch shared d1")
+            # In the absence of weight sharing (S=1), this simply computes
+            # (Q^T X_rot)^2 and (Q^T Y_rot)^2, then computes the correction
+            X_gram = einsum(X_rot, X_rot, "batch s d2, batch t d2 -> batch s t d2")
+            Y_gram = einsum(Y_rot, Y_rot, "batch s d1, batch t d1 -> batch s t d1")
+            eigencorrection = einsum(
+                Y_gram, X_gram, "batch s t d1, batch s t d2 -> d1 d2"
+            )
+
+        else:  # -> per-example gradient approach
+            rotated_per_example_gradient = einsum(
+                Q1,
+                Y,
+                X,
+                Q2,
+                "i d1, batch shared i, batch shared j, j d2 -> batch d1 d2",
+            )
+            eigencorrection = rotated_per_example_gradient.square_().sum(dim=0)
+
+    return eigencorrection
+
+
 class EKFACLinearOperator(KFACLinearOperator):
     """Linear operator to multiply with the Fisher/GGN's EKFAC approximation.
 
@@ -439,101 +641,33 @@ class EKFACLinearOperator(KFACLinearOperator):
             and "weight" in param_pos.keys()
             and "bias" in param_pos.keys()
         ):
-            activations = cat(
+            a_augmented = cat(
                 [activations, activations.new_ones(*activations.shape[:-1], 1)], dim=-1
             )
-            if g.shape[1] == 1:
-                # No weight sharing: We can take the square of the per-example gradient
-                # and sum over the batch axis more efficiently without building up the
-                # large per-example gradient tensor
-                rotated_activations = einsum(
-                    aaT_eigenvectors,
-                    activations.squeeze(1),
-                    "i d_in, batch i -> batch d_in",
-                )
-                rotated_g = einsum(
-                    ggT_eigenvectors, g.squeeze(1), "i d_out, batch i -> batch d_out"
-                )
-                eigencorrection = einsum(
-                    rotated_activations.square_(),
-                    rotated_g.square_(),
-                    "batch d_in, batch d_out -> d_out d_in",
-                )
-                self._corrected_eigenvalues = self._set_or_add_(
-                    self._corrected_eigenvalues,
-                    module_name,
-                    eigencorrection.mul_(correction),
-                )
+            eigencorrection = compute_eigenvalue_correction_linear_weight_sharing(
+                g, ggT_eigenvectors, a_augmented, aaT_eigenvectors
+            )
+            self._corrected_eigenvalues = self._set_or_add_(
+                self._corrected_eigenvalues,
+                module_name,
+                eigencorrection.mul_(correction),
+            )
 
-            else:
-                # Compute the rotated per-example gradients
-                rotated_per_example_gradient = einsum(
-                    ggT_eigenvectors,
-                    g,
-                    activations,
-                    aaT_eigenvectors,
-                    "i d_out, batch shared i, batch shared j, j d_in -> batch d_out d_in",
-                )
-                # Obtain the correction contributed by the current batch
-                self._corrected_eigenvalues = self._set_or_add_(
-                    self._corrected_eigenvalues,
-                    module_name,
-                    rotated_per_example_gradient.square_().sum(dim=0).mul_(correction),
-                )
         else:
             if module_name not in self._corrected_eigenvalues:
                 self._corrected_eigenvalues[module_name] = {}
             for p_name, pos in param_pos.items():
-                # Compute the rotated per-example gradients
-                if p_name == "weight" and g.shape[1] == 1:
-                    # No weight sharing: We can take the square of the per-example gradient
-                    # and sum over the batch axis more efficiently without building up the
-                    # large per-example gradient tensor
-                    rotated_activations = einsum(
-                        aaT_eigenvectors,
-                        activations.squeeze(1),
-                        "i d_in, batch i -> batch d_in",
-                    )
-                    rotated_g = einsum(
-                        ggT_eigenvectors,
-                        g.squeeze(1),
-                        "i d_out, batch i -> batch d_out",
-                    )
-                    eigencorrection = einsum(
-                        rotated_activations.square_(),
-                        rotated_g.square_(),
-                        "batch d_in, batch d_out -> d_out d_in",
-                    )
-                    self._corrected_eigenvalues[module_name] = self._set_or_add_(
-                        self._corrected_eigenvalues[module_name],
-                        pos,
-                        eigencorrection.mul_(correction),
-                    )
-
-                else:
-                    rotated_per_example_gradient = (
-                        einsum(
-                            ggT_eigenvectors,
-                            g,
-                            activations,
-                            aaT_eigenvectors,
-                            "i d_out, batch shared i, batch shared j, j d_in -> batch d_out d_in",
-                        )
-                        if p_name == "weight"
-                        else einsum(
-                            ggT_eigenvectors,
-                            g,
-                            "j d_out, batch shared j -> batch d_out",
-                        )
-                    )
-                    # Obtain the correction contributed by the current batch
-                    self._corrected_eigenvalues[module_name] = self._set_or_add_(
-                        self._corrected_eigenvalues[module_name],
-                        pos,
-                        rotated_per_example_gradient.square_()
-                        .sum(dim=0)
-                        .mul_(correction),
-                    )
+                eigencorrection = compute_eigenvalue_correction_linear_weight_sharing(
+                    g,
+                    ggT_eigenvectors,
+                    activations,
+                    aaT_eigvecs=None if p_name == "bias" else aaT_eigenvectors,
+                )
+                self._corrected_eigenvalues[module_name] = self._set_or_add_(
+                    self._corrected_eigenvalues[module_name],
+                    pos,
+                    eigencorrection.mul_(correction),
+                )
 
     @property
     def trace(self) -> Tensor: