+++ changelog.md
@@ -15,13 +15,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add EKFAC and a new problem (ResNet18 on CIFAR10) to benchmark
   ([PR](https://github.com/f-dangel/curvlinops/pull/214)),
   expose a run time and memory inefficiency in EKFAC
-  ([issue](https://github.com/f-dangel/curvlinops/issues/193))
+  ([issue](https://github.com/f-dangel/curvlinops/issues/193)).
+  This issue is successfully resolved with:
 
   - Improve performance of EKFAC by merging gradient computation and basis rotation into a single `einsum`
   ([PR](https://github.com/f-dangel/curvlinops/pull/215))
 
   - Improve performance of EKFAC by efficiently taking the square and sum over the batch dimension in the absence of weight sharing
   ([PR](https://github.com/f-dangel/curvlinops/pull/216))
+
+  - Generalize trick for no weight sharing to mild weight sharing, further improve memory performance of EKFAC
+  ([PR](https://github.com/f-dangel/curvlinops/pull/219))
 
 - Minor fixes in docs, code and examples to reduce CI pipeline errors ([PR](https://github.com/f-dangel/curvlinops/pull/218))